Extract classicEditorData functions to helpers and implement in data.js.
