+++ js/src/analysis/classicEditorData.js
@@ -1,11 +1,13 @@
 /* External dependencies */
 import removeMarks from "yoastseo/js/markers/removeMarks";
-import forEach from "lodash/forEach";
-import omit from "lodash/omit";
 
 /* Internal dependencies */
 import { updateReplacementVariable } from "../redux/actions/snippetEditor";
-import updateReplacementVariables from "../helpers/updateReplacementVariables";
+import {
+	fillReplacementVariables,
+	mapCustomFields,
+	mapCustomTaxonomies,
+} from "../helpers/replacementVariableHelpers";
 import tmceHelper, { tmceId } from "../wp-seo-tinymce";
 import debounce from "lodash/debounce";
 
@@ -39,7 +41,7 @@ class ClassicEditorData {
 	 */
 	initialize( replaceVars ) {
 		this._data = this.getInitialData( replaceVars );
-		updateReplacementVariables( this._data, this._store );
+		fillReplacementVariables( this._data, this._store );
 		this.subscribeToElements();
 		this.subscribeToStore();
 	}
@@ -195,53 +197,6 @@ class ClassicEditorData {
 	}
 
 	/**
-	 * Map the custom_fields field in the replacevars to a format suited for redux.
-	 *
-	 * @param {Object} replaceVars       The original replacevars.
-	 *
-	 * @returns {Object}                 The restructured replacevars object without custom_fields.
-	 */
-	mapCustomFields( replaceVars ) {
-		if( ! replaceVars.custom_fields ) {
-			return replaceVars;
-		}
-
-		let customFieldReplaceVars = {};
-		forEach( replaceVars.custom_fields, ( value, key ) => {
-			customFieldReplaceVars[ `cf_${ key }` ] = value;
-		} );
-
-		return omit( {
-			...replaceVars,
-			...customFieldReplaceVars,
-		}, "custom_fields" );
-	}
-
-	/**
-	 * Map the custom_taxonomies field in the replacevars to a format suited for redux.
-	 *
-	 * @param {Object} replaceVars       The original replacevars.
-	 *
-	 * @returns {Object}                 The restructured replacevars object without custom_taxonomies.
-	 */
-	mapCustomTaxonomies( replaceVars ) {
-		if( ! replaceVars.custom_taxonomies ) {
-			return replaceVars;
-		}
-
-		let customTaxonomyReplaceVars = {};
-		forEach( replaceVars.custom_taxonomies, ( value, key ) => {
-			customTaxonomyReplaceVars[ `ct_${ key }` ] = value.name;
-			customTaxonomyReplaceVars[ `ct_desc_${ key }` ] = value.description;
-		} );
-
-		return omit( {
-			...replaceVars,
-			...customTaxonomyReplaceVars,
-		}, "custom_taxonomies" );
-	}
-
-	/**
 	 * Gets the initial data from the replacevars and document.
 	 *
 	 * @param {Object} replaceVars The replaceVars object.
@@ -249,8 +204,8 @@ class ClassicEditorData {
 	 * @returns {Object} The data.
 	 */
 	getInitialData( replaceVars ) {
-		replaceVars = this.mapCustomFields( replaceVars );
-		replaceVars = this.mapCustomTaxonomies( replaceVars );
+		replaceVars = mapCustomFields( replaceVars );
+		replaceVars = mapCustomTaxonomies( replaceVars );
 		return {
 			...replaceVars,
 			title: this.getTitle(),
