--- js/src/helpers/updateReplacementVariables.js
@@ -1,21 +1,0 @@
-import forEach from "lodash/forEach";
-import { updateReplacementVariable } from "../redux/actions/snippetEditor";
-
-export const nonReplaceVars = [ "slug", "content" ];
-
-/**
- * Fills the redux store with the newly acquired data.
- *
- * @param {Object} data  The data object.
- * @param {Object} store The redux store.
- *
- * @returns {void}
- */
-export default function updateReplacementVariables( data, store ) {
-	forEach( data, ( value, name ) => {
-		if ( nonReplaceVars.includes( name ) ) {
-			return;
-		}
-		store.dispatch( updateReplacementVariable( name, value ) );
-	} );
-}
